requirements.txt
Package        Version
-------------- -------
pip            20.0.2 
pkg-resources  0.0.0  
pytz           2020.1 
PyYAML         5.3.1  
setuptools     44.0.0 
wheel          0.35.1 
zope.interface 5.1.2  
